Fix two serious routing bugs
The `dig4_c` trace was, unfortunately, left unconnected after my night
pour optimization (...). I had to add a fixup wire to the first
prototype.

Also, the Pi module's footprint is a bit larger than the silkscreen
because the SD card protrudes a little bit to the left, towards the
rotary encoder. I tried to make this less urgent (note that I'm
deliberately not saying "fixed") by moving the rotary encoder a bit to
the left, which necessitated moving the PWM testpoints as well, which
required some rerouting. Meh.

I think that a real fix is to leave the SD card accessible, which will
mean rotating the Pi module 180 degrees. Some other time, perhaps.